+++ Doctrine/Phpcr/StaticContent.php
@@ -42,22 +42,6 @@ class StaticContent extends ModelStaticContent implements HierarchyInterface
     protected $node;
 
     /**
-     * @deprecated Use setParentDocument instead.
-     */
-    public function setParent($parent)
-    {
-        $this->setParentDocument($parent);
-    }
-
-    /**
-     * @deprecated Use getParentDocument instead.
-     */
-    public function getParent()
-    {
-        return $this->getParentDocument();
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function setParentDocument($parent)